week22_streaming_processing/pyproject.toml
[project]
name = "week22-streaming-processing"
version = "0.1.0"
description = ""
authors = [
    {name = "Emma Berniczky",email = "e.berniczky@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-pubsub (>=2.30.0,<3.0.0)",
    "apache-beam (>=2.65.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
